https://aktdiverxamotos.com/
Hello team,

I currently work in a fraud company in which a fake site of  motos AKT company, has been detected.

The site uses the official image of the company, the name and displays a website similar to the official one. The fake site was verified and may collect user data and credit card information in exchange for sales.

The fake site (https://aktdiverxamotos.com/) has been reported to the company. They confirm that their official domain is (https://www.aktmotos.com/)


I appreciate your help reporting the site to disable it.