Phishing
 Hello Team;

I currently work in a fraud company and fake sites of  Alkosto, has been detected.
 

The sites uses the official image of the company, the name and displays a website similar to the official one. The fake sites was verified and may collect user data and credit card information in exchange for sales. 

The fake sites liste are:

https://bluntunlock.top/alkosto/
https://compatibletimber.xyz/B1huMFgy/alkosto/
https://eternallesson.xyz/exito/
https://varioususage.top/alkosto/
https://jollyoblige.top/alkosto/
https://talentll.top/3U6KPWso/alkosto/

The list has been reported to the company. They confirm that their official domain is (https://www.alkosto.com/)
 

I appreciate your help reporting the site to disable it.