Update add-link
Hello Team;

I currently work in a fraud company and a fake site of  Alkosto.com, has been detected. 

The site uses the official image of the company, the name and displays a website similar to the official one. The fake site was verified and may collect user data and credit card information in exchange for sales.

Te website identifies is:  https://usualuntil.top/alkosto/

The site has been reported to the company. They confirm that their official domain is (https://www.alkosto.com/)
 
I appreciate your help reporting the site to disable it.