Update add-link
Hello Team;

I currently work in a fraud company and fake sites of  Alkosto.com, has been detected.
 

The sites use the official image of the company, the name and displays a website similar to the official one. The fake sites was verified and may collect user data and credit card information in exchange for sales.

The sites identified are:

https://recallvapor.top/alkosto/
https://benignvan.xyz/alkosto/

The sites has been reported to the company. They confirm that their official domain is (https://www.alkosto.com/)
 

I appreciate your help reporting the site to disable it.